Create standalone app for local testing
